Converted availability-set tests
