--- test/d2l-organization-availability-set/d2l-current-organization-availability.html
@@ -1,20 +1,0 @@
-<html>
-	<head>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes">
-		<script src="../../../@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
-		<script src="../../../wct-browser-legacy/browser.js"></script>
-
-		<script type="module" src="../../../siren-parser/global.js"></script>
-		<script type="module" src="../../components/d2l-organization-availability-set/d2l-current-organization-availability.js"></script>
-	</head>
-	<body>
-		<test-fixture id="current-org-availability">
-			<template>
-				<d2l-current-organization-availability token="whatever"></d2l-current-organization-availability>
-			</template>
-		</test-fixture>
-
-		<script type="module" src="./d2l-current-organization-availability.js"></script>
-	</body>
-</html>
